# contracts/openapi/zen70-gateway-full.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZEN70 API",
    "version": "1.58.0"
  },
  "paths": {
...
    "/api/v1/auth/webauthn/register/begin": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register Begin",
        "operationId": "register_begin_api_v1_auth_webauthn_register_begin_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebAuthnRegisterBeginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebAuthnRegisterBeginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/webauthn/register/complete": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register Complete",
        "operationId": "register_complete_api_v1_auth_webauthn_register_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebAuthnRegisterCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Register Complete Api V1 Auth Webauthn Register Complete Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/webauthn/login/begin": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login Begin",
        "operationId": "login_begin_api_v1_auth_webauthn_login_begin_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebAuthnLoginBeginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebAuthnLoginBeginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/webauthn/login/complete": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login Complete",
        "operationId": "login_complete_api_v1_auth_webauthn_login_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebAuthnLoginCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "description": "JWT access token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "description": "过期秒数"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_in"
        ],
        "title": "TokenResponse",
        "description": "令牌响应。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WebAuthnLoginBeginRequest": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Tenant Id",
            "default": "default"
          },
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Username"
          }
        },
        "type": "object",
        "required": [
          "username"
        ],
        "title": "WebAuthnLoginBeginRequest",
        "description": "WebAuthn 登录开始请求。"
      },
      "WebAuthnLoginBeginResponse": {
        "properties": {
          "options": {
            "additionalProperties": true,
            "type": "object",
            "title": "Options",
            "description": "JSON 格式的 PublicKeyCredentialRequestOptions"
          }
        },
        "type": "object",
        "required": [
          "options"
        ],
        "title": "WebAuthnLoginBeginResponse",
        "description": "WebAuthn 登录开始响应。"
      },
      "WebAuthnLoginCompleteRequest": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Tenant Id",
            "default": "default"
          },
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Username"
          },
          "credential": {
            "additionalProperties": true,
            "type": "object",
            "title": "Credential",
            "description": "navigator.credentials.get() 的返回值"
          }
        },
        "type": "object",
        "required": [
          "username",
          "credential"
        ],
        "title": "WebAuthnLoginCompleteRequest",
        "description": "WebAuthn 登录完成请求。"
      },
      "WebAuthnRegisterBeginRequest": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Tenant Id",
            "default": "default"
          },
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Username"
          },
          "display_name": {
            "type": "string",
            "maxLength": 128,
            "title": "Display Name",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "username"
        ],
        "title": "WebAuthnRegisterBeginRequest",
        "description": "WebAuthn 注册开始请求。"
      },
      "WebAuthnRegisterBeginResponse": {
        "properties": {
          "options": {
            "additionalProperties": true,
            "type": "object",
            "title": "Options",
            "description": "JSON 格式的 PublicKeyCredentialCreationOptions"
          }
        },
        "type": "object",
        "required": [
          "options"
        ],
        "title": "WebAuthnRegisterBeginResponse",
        "description": "WebAuthn 注册开始响应：options 可直接传给 navigator.credentials.create()。"
      },
      "WebAuthnRegisterCompleteRequest": {
        "properties": {
          "credential": {
            "additionalProperties": true,
            "type": "object",
            "title": "Credential",
            "description": "navigator.credentials.create() 的返回值"
          }
        },
        "type": "object",
        "required": [
          "credential"
        ],
        "title": "WebAuthnRegisterCompleteRequest",
        "description": "WebAuthn 注册完成请求：前端传来的 credential 对象。"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}